betting-chatbot/requirements.txt
# Core dependencies
Flask==2.3.3
Flask-Cors==4.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
Werkzeug==2.3.7

# Authentication
PyJWT==2.8.0
bcrypt==4.0.1

# Database
SQLAlchemy==2.0.23

# OpenAI API
openai==1.3.5

# Testing
pytest==7.4.3
pytest-flask==1.3.0

# Utilities
requests==2.31.0
python-dateutil==2.8.2
